Move address range helpers into address package
This means that users don't drag in a dependency on aws-sdk.
